+++ CHANGELOG.md
@@ -6,7 +6,17 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## Unreleased
 
-* Nextaur updated to 1.7.0-beta on staging
+### Added
+
+* certifi dxpy dependency
+
+### Changed
+
+* urllib3 dxpy dependency to >=1.26.18,<2.2
+
+### Removed
+
+* requests, cryptography as dependencies of dxpy
 
 ## [365.0] - beta